Commands refactor
